requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
langchain_google_genai
pathlib
langchain
faiss-cpu==1.7.4
langchain-core
langchain-text-splitters
sentence-transformers